Make iscsi.global platform specific
